requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
catboost
python-dotenv>=0.5.1
